requirements.txt
google-genai>=0.1.0
python-docx>=1.1.0
reportlab>=4.1.0
pydantic>=2.6.0
python-dotenv>=1.0.1
